+++ src/pc.hpp
@@ -11,6 +11,7 @@
 #include <vector>
 #include <sstream>
 #include <unordered_map>
+#include <optional>
 
 #include "robin_hood.h"
 #include "ssw_cpp.h"
@@ -23,19 +24,28 @@ class InputBuffer {
 
 public:
 
-    InputBuffer(input_stream_t& ks1, input_stream_t& ks2, int chunk_size)
-    : ks1(ks1), ks2(ks2), chunk_size(chunk_size) { }
+    InputBuffer(std::string fname1, std::string fname2, int chunk_size, bool is_interleaved)
+    : ks1(open_fastq(fname1)),
+    ks2(fname2 == "" ? nullptr : open_fastq(fname2)),
+    chunk_size(chunk_size),
+    is_interleaved(is_interleaved) { }
 
     std::mutex mtx;
 
-    input_stream_t &ks1;
-    input_stream_t &ks2;
+    input_stream_t ks1;
+    input_stream_t ks2;
+    std::optional<klibpp::KSeq> lookahead1;
     bool finished_reading{false};
     int chunk_size;
     size_t chunk_index{0};
+    bool is_interleaved{false};
 
-    size_t read_records_PE(std::vector<klibpp::KSeq> &records1, std::vector<klibpp::KSeq> &records2, AlignmentStatistics &statistics);
-    size_t read_records_SE(std::vector<klibpp::KSeq> &records1, AlignmentStatistics &statistics);
+    void rewind_reset();
+    size_t read_records(std::vector<klibpp::KSeq> &records1,
+            std::vector<klibpp::KSeq> &records2,
+            std::vector<klibpp::KSeq> &records3,
+            AlignmentStatistics &statistics,
+            int read_count=-1);
 };
 
 
@@ -53,12 +63,8 @@ public:
 };
 
 
-void perform_task_PE(InputBuffer &input_buffer, OutputBuffer &output_buffer,
+void perform_task(InputBuffer &input_buffer, OutputBuffer &output_buffer,
                   AlignmentStatistics& statistics, const alignment_params &aln_params,
                   const mapping_params &map_param, const IndexParameters& index_parameters, const References& references, const StrobemerIndex& index, const std::string& read_group_id);
-
-void perform_task_SE(InputBuffer &input_buffer, OutputBuffer &output_buffer,
-                     AlignmentStatistics& statistics, const alignment_params &aln_params,
-                     const mapping_params &map_param, const IndexParameters& index_parameters, const References& references, const StrobemerIndex& index, const std::string& read_group_id);
 
 #endif